api can now take post and get request
In this update, the flask rest api can now handle post request that stores data from the client side and store it locally and temporary. Then, the server will return the url corresponded to the data to the client which they can use to do a GET request and view the qr code that was generated based on their data. Some example qr code .png files are also included.